requirements.txt
Pillow==9.1.1
numpy==1.22.4
